% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_data.R
\docType{data}
\name{nor_fullform_pos_sent_positive}
\alias{nor_fullform_pos_sent_positive}
\title{Norwegian positive sentiment dictionary in fullform with part-of-speech}
\format{
An object of class \code{\link[base]{data.frame}} containing all positive fullform tokens
and their parts-of speech over three variables:\tabular{ll}{
    \tab  \cr
   \strong{token} \tab The token (word) in fullform \cr
   \strong{pos} \tab Part-of-speech for the token \cr
   \strong{is_base_form} \tab Whether the token is the base form of the word ("yes") or not ("no") \cr
}
}
\usage{
data(nor_fullform_pos_sent_positive)
}
\description{
Data containting sentiment dictionaries for positive fullform
tokens with part-of-speech tags in Norwegian, based on the
\emph{Sentiment Analysis for Norwegian Text} project at the Department
of Informatics, University of Oslo.
}
\examples{
data(nor_fullform_pos_sent_positive)

head(nor_fullform_pos_sent_positive)
}
\references{
Barnes et al. (2019) Lexicon information in neural sentiment
analysis: a multi-task learning approach. Proceedings of the 22nd Nordic
Conference on Computational Linguistics. Turku, Finland
(\href{https://www.aclweb.org/anthology/W19-6119/}{ACL Anthology})
}
\keyword{datasets}
