% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_lemma_sent.R
\docType{data}
\name{nor_lemma_sent}
\alias{nor_lemma_sent}
\title{Norwegian sentiment dictionary in fullform}
\format{
An object of class \code{\link[base]{list}} containing one character
vector for positive and negative lemma tokens across different types of
parts-of-speech
}
\usage{
data(nor_lemma_sent)
}
\description{
Data containting sentiment dictionaries for positive and negative lemma
tokens in Norwegian, based on the _Sentiment Analysis for Norwegian Text_
project at the Department of Informatics, University of Oslo.
}
\examples{
data(nor_lemma_sent)

sent_pos <- nor_lemma_sent$lemma_noun_positive
sent_neg <- nor_lemma_sent$lemma_noun_negative

head(sent_pos)
head(sent_neg)
}
\references{
Barnes et al. (2019) Lexicon information in neural sentiment
analysis: a multi-task learning approach. Proceedings of the 22nd Nordic
Conference on Computational Linguistics. Turku, Finland
(\href{https://www.aclweb.org/anthology/W19-6119/}{ACL Anthology})
}
\keyword{datasets}
